Data binding

+++ app/src/main/java/com/example/shoppinglist/domain/ShopItem.kt
@@ -2,7 +2,7 @@ package com.example.shoppinglist.domain
 
 data class ShopItem(
     val name: String,
-    val weight: Float,
+    val weight: Double,
     val count: Int,
     val enabled: Boolean,
     var id: Int = UNDEFINED_ID,
